Quote reverser api model refactoring
There was a model.go that contains model named InputQuote
for reading quotes and OutputQuote for each quote array
grouped by author. And there was a map key as an author and
value as an OutputQuote. These objects names and functionality
was wrong. Map were containing author name 2 times for each element.
It is fixed. Now there is a struct Quote that represents for each read
quote from api. AuthorQuotes struct that represent for each quotes array
grouped by author. AuthorQuotesMap is a map uses author name as a key and
quotes of this author as a value.